requirements.txt
Django==4.1.4
django-timezone-field==5.0
djangorestframework==3.14.0
jugaad-data==0.24
numpy==1.24.0
django-cors-headers==3.12.0
pandas==1.5.2
requests==2.28.1
simplejson==3.18.0
sqlparse==0.4.3
websockets==10.4
zipp==3.11.0
dj-static==0.0.6
whitenoise==6.3.0   